+++ Fridges.Application/Services/RecipeService.cs
@@ -1,5 +1,7 @@
+using System.Collections.Generic;
 using Fridges.Application.DTOs;
 using Fridges.Application.Interfaces;
+using Fridges.Domain.Enities.Products;
 using Fridges.Domain.Enities.Recipe;
 using Fridges.Persistance.Interfaces;
 
@@ -13,20 +15,44 @@ public class RecipeService : IRecipeService
         _recipeRepository = recipeRepository;
     }
 
-    public async Task<List<Recipe>> GetAllAsync()
+    public async Task<Result<List<Recipe>>> GetAllAsync()
     {
         var recipes = await _recipeRepository.GetAllAsync();
-        return recipes;
+        if (recipes == null)
+        {
+            return Result<List<Recipe>>.Failure("No recipes found");
+        }
+        else
+        {
+            return Result<List<Recipe>>.Success(recipes);
+        }
     }
 
-    public async Task<Recipe> GetAsync(Guid id)
+    public async Task<Result<Recipe>> GetAsync(Guid id)
     {
         var recipe = await _recipeRepository.GetByIdAsync(id);
-        return recipe;
+        if( recipe == null)
+        {
+            return Result<Recipe>.Failure($"Fridge with {id} not found");
+        }
+        return Result<Recipe>.Success(recipe);
     }
 
-    public async Task<Recipe> AddAsync(RecipeDto dto)
+    public async Task<Result<Recipe>> AddAsync(RecipeDto dto)
     {
+        if (string.IsNullOrWhiteSpace(dto.Instructions))
+        {
+             return Result<Recipe>.Failure("Instructions field should be filled");
+        }
+        if (string.IsNullOrWhiteSpace(dto.title))
+        {
+            return Result<Recipe>.Failure("Title field is required");
+        }
+        if (dto.Products == null)
+        {
+            return Result<Recipe>.Failure("Recipe should have products");
+        }
+
         var recipe = new Recipe
         {
             Id = Guid.NewGuid(),
@@ -40,17 +66,41 @@ public class RecipeService : IRecipeService
             }).ToList()
         };
         await _recipeRepository.AddAsync(recipe);
-        return recipe;
+
+        return Result<Recipe>.Success(recipe);
     }
 
-    public async Task DeleteAsync(Guid id)
-    {
-        await _recipeRepository.DeleteAsync(id);
-    }
-
-    public async Task EditAsync(Guid id, RecipeDto dto)
+    public async Task<Result> DeleteAsync(Guid id)
     {
         var recipe = await _recipeRepository.GetByIdAsync(id);
+        if(recipe == null)
+        {
+            return Result.Failure($"Recipe with {id} doesn't exist");
+        }
+        await _recipeRepository.DeleteAsync(id);
+        return Result.Success();
+    }
+
+    public async Task<Result<Recipe>> EditAsync(Guid id, RecipeDto dto)
+    {
+        var recipe = await _recipeRepository.GetByIdAsync(id);
+        if (recipe == null)
+        {
+            return Result<Recipe>.Failure($"Recipe with {id} doesn't exist");
+        }
+        if (string.IsNullOrWhiteSpace(dto.Instructions))
+        {
+            return Result<Recipe>.Failure("Instructions field should be filled");
+        }
+        if (string.IsNullOrWhiteSpace(dto.title))
+        {
+            return Result<Recipe>.Failure("Title field is required");
+        }
+        if (dto.Products == null)
+        {
+            return Result<Recipe>.Failure("Recipe should have products");
+        }
+
         recipe.title = dto.title;
         recipe.Instructions = dto.Instructions;
         List<RecipeProduct> products = dto.Products
@@ -65,11 +115,17 @@ public class RecipeService : IRecipeService
 
         await _recipeRepository.CheckRecipeProductsAsync(id, products);
         await _recipeRepository.UpdateAsync(recipe);
+        return Result<Recipe>.Success(recipe);
     }
 
-    public async Task<List<RecipeProduct>> GetRecipeProductsByIdAsync(Guid id)
+    public async Task<Result<List<RecipeProduct>>> GetRecipeProductsByIdAsync(Guid id)
     {
+        var recipe = await _recipeRepository.GetByIdAsync(id);
+        if (recipe == null)
+        {
+            return Result<List<RecipeProduct>>.Failure($"Recipe with {id} doesn't exist");
+        }
         var products = await _recipeRepository.GetRecipeProductsByIdAsync(id);
-        return products;
+        return Result<List<RecipeProduct>>.Success(products);
     }
 }
